+++ app/Http/Controllers/UserPrivilegeController.php
@@ -4,11 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\UserPrivilege;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Validator;
 
 class UserPrivilegeController extends Controller
 {
+    const blade_view = 'layout.user.privilege';
+    const url_redirect ='user/privilege';
+    const name = 'user privilege';
+    const privilege = 'users';
+
     /**
      * Display a listing of the resource.
      *
@@ -17,16 +21,16 @@ class UserPrivilegeController extends Controller
     public function index()
     {
         // cek privilege
-        privilegeLevel('users', ONLY_SEE);
+        privilegeLevel(self::privilege, ONLY_SEE);
 
         // penguraian data
         $params = [
             'data'  => UserPrivilege::all(),
             'type'  => 'data',
-            'title' => 'User Privilege'
+            'title' => self::name
         ];
 
-        return view('layout.user.privilege.data', $params);
+        return view(self::blade_view.'.data', $params);
     }
 
     /**
@@ -37,7 +41,7 @@ class UserPrivilegeController extends Controller
     public function create()
     {
         // cek privilege
-        privilegeLevel('users', CAN_CRUD);
+        privilegeLevel(self::privilege, CAN_CRUD);
 
         $data = [
             'id'        => null,
@@ -57,10 +61,10 @@ class UserPrivilegeController extends Controller
         $params = [
             'data'  => $data,
             'type'  => 'create',
-            'title' => 'Create User Privilege'
+            'title' => 'Create '.self::name
         ];
 
-        return view('layout.user.privilege.input', $params);
+        return view(self::blade_view.'.input', $params);
     }
 
     /**
@@ -72,7 +76,7 @@ class UserPrivilegeController extends Controller
     public function store(Request $request)
     {
         // cek privilege
-        privilegeLevel('users', CAN_CRUD);
+        privilegeLevel(self::privilege, CAN_CRUD);
 
         // validasi
         $rules = [
@@ -94,20 +98,10 @@ class UserPrivilegeController extends Controller
 
         $hasil = UserPrivilege::create($request->all());
 
-        if($hasil){
-            $params = [
-                'status'    => 'success',
-                'message'   => 'Sukses menambahkan user privilege'
-            ];
-        } else {
-            $params = [
-                'status'    => 'error',
-                'message'   => 'Gagal menambahkan user privilege'
-            ];
-        }
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'create', self::name);
 
-//        AddLog::add('<b>'.ucwords(Auth::user()->name).'</b> telah '.strtolower($params['message']), 'privilege', $params['status'] == 'success' ? $privilege->id : null);
-        return redirect('user/privilege')->with($params);
+        return redirect(self::url_redirect)->with($params);
     }
 
     /**
@@ -130,16 +124,16 @@ class UserPrivilegeController extends Controller
     public function edit(UserPrivilege $userPrivilege, $id)
     {
         // cek privilege
-        privilegeLevel('users', CAN_CRUD);
+        privilegeLevel(self::privilege, CAN_CRUD);
 
         // penguraian data
         $params = [
             'data'  => $userPrivilege->find($id),
             'type'  => 'edit',
-            'title' => 'Edit Data User Privilege'
+            'title' => 'Edit Data '.self::name
         ];
 
-        return view('layout.user.privilege.input', $params);
+        return view(self::blade_view.'.input', $params);
     }
 
     /**
@@ -152,7 +146,7 @@ class UserPrivilegeController extends Controller
     public function update(Request $request, UserPrivilege $userPrivilege, $id)
     {
         // cek privilege
-        privilegeLevel('users', CAN_CRUD);
+        privilegeLevel(self::privilege, CAN_CRUD);
 
         // validasi
         $rules = [
@@ -175,20 +169,10 @@ class UserPrivilegeController extends Controller
         $userPrivilege = $userPrivilege->find($id);
         $hasil = $userPrivilege->fill($request->all())->save();
 
-        if($hasil){
-            $params = [
-                'status'    => 'success',
-                'message'   => 'Sukses mengubah user privilege'
-            ];
-        } else {
-            $params = [
-                'status'    => 'error',
-                'message'   => 'Gagal mengubah user privilege'
-            ];
-        }
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'update', self::name);
 
-//        AddLog::add('<b>'.ucwords(Auth::user()->name).'</b> telah '.strtolower($params['message']), 'privilege', $id ? $id : null);
-        return redirect('user/privilege')->with($params);
+        return redirect(self::url_redirect)->with($params);
     }
 
     /**
@@ -200,31 +184,21 @@ class UserPrivilegeController extends Controller
     public function destroy(UserPrivilege $userPrivilege, $id)
     {
         // cek privilege
-        privilegeLevel('users', CAN_CRUD);
+        privilegeLevel(self::privilege, CAN_CRUD);
 
         $hasil = UserPrivilege::find($id);
         $hasil->delete();
 
-        if($hasil){
-            $params = [
-                'status'    => 'success',
-                'message'   => 'Sukses menghapus user privilege'
-            ];
-        } else {
-            $params = [
-                'status'    => 'error',
-                'message'   => 'Gagal menghapus user privilege'
-            ];
-        }
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'delete', self::name);
 
-//        AddLog::add('<b>'.ucwords(Auth::user()->name).'</b> telah '.strtolower($params['message']), 'privilege', $id ? $id : null);
-        return redirect('user/privilege')->with($params);
+        return redirect(self::url_redirect)->with($params);
     }
 
     public function trash()
     {
         // cek privilege
-        privilegeLevel('users', ALL_ACCESS);
+        privilegeLevel(self::privilege, ALL_ACCESS);
 
         // penguraian data
         $params = [
@@ -233,13 +207,13 @@ class UserPrivilegeController extends Controller
             'title' => 'Trash'
         ];
 
-        return view('layout.user.privilege.data', $params);
+        return view(self::blade_view.'.data', $params);
     }
 
     public function restore($id = null)
     {
         // cek privilege
-        privilegeLevel('users', ALL_ACCESS);
+        privilegeLevel(self::privilege, ALL_ACCESS);
 
         if ($id != null){
             $hasil = UserPrivilege::onlyTrashed()
@@ -249,25 +223,16 @@ class UserPrivilegeController extends Controller
             $hasil = UserPrivilege::onlyTrashed()->restore();
         }
 
-        if($hasil){
-            $params = [
-                'status'    => 'success',
-                'message'   => 'Sukses mengembalikan user privilege'
-            ];
-        } else {
-            $params = [
-                'status'    => 'error',
-                'message'   => 'Gagal mengembalikan user privilege'
-            ];
-        }
-//        AddLog::add('<b>'.ucwords(Auth::user()->name).'</b> telah '.strtolower($params['message']), 'user', $id ? $id : null);
-        return redirect('user/privilege/trash')->with($params);
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'restore', self::name);
+
+        return redirect(self::url_redirect.'/trash')->with($params);
     }
 
     public function delete($id = null)
     {
         // cek privilege
-        privilegeLevel('users', ALL_ACCESS);
+        privilegeLevel(self::privilege, ALL_ACCESS);
 
         if ($id != null){
             $hasil = UserPrivilege::onlyTrashed()
@@ -277,18 +242,93 @@ class UserPrivilegeController extends Controller
             $hasil = UserPrivilege::onlyTrashed()->forceDelete();
         }
 
-        if($hasil){
-            $params = [
-                'status'    => 'success',
-                'message'   => 'Sukses menghapus permanen user privilege'
-            ];
-        } else {
-            $params = [
-                'status'    => 'error',
-                'message'   => 'Gagal menghapus permanen user privilege'
-            ];
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'delete permanent', self::name);
+
+        return redirect(self::url_redirect.'/trash')->with($params);
+    }
+
+    public function get_table(Request $request)
+    {
+        /**
+         * Parameters
+         */
+        $start = $request->start ? intval($request->start) : 0;
+        $length = $request->length ? intval($request->length) : 10;
+        $search = empty($request->searchQuery) || $request->searchQuery === "null" ? "" : $request->searchQuery;
+        $sortColumnIndex = $request->sortColumn;
+        $sortDirection = $request->sortDirection;
+        $type = $request->type;
+
+        /**
+         * Raw Data
+         */
+
+        if ($type == 'data') {
+            $list = UserPrivilege::select('*');
         }
-//        AddLog::add('<b>'.ucwords(Auth::user()->name).'</b> telah '.strtolower($params['message']), 'user', $id ? $id : null);
-        return redirect('user/privilege/trash')->with($params);
+        if ($type == 'trash') {
+            $list = UserPrivilege::onlyTrashed();
+        }
+
+        // search
+        if ($search) {
+            $list = $list->where('name', 'like', '%'.$search.'%')
+                ->orWhere('color', 'like', '%'.$search.'%');
+        }
+        $list = $list->get();
+
+        $data = array();
+
+        $no = $start;
+        foreach ($list as $item) {
+            $no++;
+            $row = array();
+            $row[] = [
+                'data'          => $no,
+                'attributes'    => [
+                    'class'     => 'text-center text-muted w-1-slot'
+                ]
+            ];
+            $row[] = $item->name;
+            $row[] = $item->color ? '<span class="'.getBadge($item->color).'">'.ucwords($item->color).'</span>' : '<span>Unset</span>';
+            if (getUserLevel('users') >= CAN_CRUD) {
+                $row[] =  [
+                    'data'          => '<div class="d-flex justify-content-center">
+                            '.getButtonCRUD($type, self::url_redirect, $item->id).'
+                          </div>',
+                    'attributes'    => [
+                        'class'     => 'w-2-slot'
+                    ]
+                ];
+            };
+            $data[] = $row;
+        }
+
+//        $data = json_decode(UserPrivilege::all()->toJson());
+
+        /**
+         * Search
+         */
+        $filtered = $data;
+
+        /*
+         * Sort
+         */
+        if (!is_null($sortColumnIndex) && $sortColumnIndex !== FALSE && $sortColumnIndex !== "null" && is_numeric($sortColumnIndex)) {
+            array_multisort(array_column($filtered, $sortColumnIndex), ($sortDirection === "asc" ? SORT_ASC : SORT_DESC), $filtered);
+        }
+
+
+        /**
+         * Slice
+         */
+        $response = array_slice($filtered, $start, $length);
+
+        echo json_encode([
+            "recordsTotal" => count($data),
+            "recordsFiltered" => count($filtered),
+            "data" => $response
+        ]);
     }
 }