+++ resources/views/layout/branch/coordinator/data.blade.php
@@ -3,7 +3,7 @@
 @section('title', ucwords($title))
 
 @section('sidebar')
-    @include('layout.user.sidebar')
+    @include('layout.branch.coordinator.sidebar')
 @endsection
 
 @section('content')
@@ -23,21 +23,21 @@
                     @endif
                     <div>
                         <h4 class="fw-bold mb-0">{{ ucwords($type) }}</h4>
-                        <small class="text-muted">Daftar list pengguna atau pekerja</small>
+                        <small class="text-muted">Daftar list koordinator</small>
                     </div>
                     {{--                Jika can CRUD maka munculkan tombol--}}
-                    @if(getUserLevel('users') >= CAN_CRUD)
+                    @if(getUserLevel('branches') >= CAN_CRUD)
                         <div class="align-self-center ms-auto">
                             @if($type == 'data')
-                                <a href="{{ url('user/create') }}" class="btn btn-primary">
+                                <a href="{{ url('branch/coordinator/create') }}" class="btn btn-primary">
                                     @svg('heroicon-s-plus', 'icon') Create new
                                 </a>
                             @endif
                             @if($type == 'trash')
-                                <a href="{{ url('user/restore') }}" class="btn btn-info me-2 restore-action-link">
+                                <a href="{{ url('branch/coordinator/restore') }}" class="btn btn-info me-2 restore-action-link">
                                     @svg('heroicon-s-arrow-up-on-square-stack', 'icon') Restore All
                                 </a>
-                                <a href="{{ url('user/delete') }}" class="btn btn-danger delete-action-link">
+                                <a href="{{ url('branch/coordinator/delete') }}" class="btn btn-danger delete-action-link">
                                     @svg('heroicon-s-trash', 'icon') Destroy All
                                 </a>
                             @endif
@@ -49,15 +49,10 @@
                         <thead>
                         <tr>
                             <th class="ps-3">#</th>
-                            <th>Image</th>
                             <th>Name</th>
-                            <th>Username</th>
-                            <th>Address</th>
-                            <th>Email</th>
-                            <th>Phone</th>
-                            <th>Privilege</th>
+                            <th>Coordinator</th>
                             {{--                                Jika can CRUD maka munculkan tombol--}}
-                            @if(getUserLevel('users') >= CAN_CRUD)
+                            @if(getUserLevel('branches') >= CAN_CRUD)
                                 <th class="pe-3 text-center">Action</th>
                             @endif
                         </tr>
@@ -67,28 +62,20 @@
                             @foreach($data as $row)
                                 <tr class="align-middle">
                                     <td class="ps-3 text-muted w-1-slot">{{ $loop->iteration }}</td>
-                                    <td>
-                                        @if($row->image)
-                                            <img src="{{ asset('uploads/images/users/'.$row->image) }}" class="avatar rounded-circle" alt="Avatar">
-                                        @else
-                                            <img src="{{ asset('uploads/images/users/default.jpg') }}" class="avatar rounded-circle" alt="Avatar">
-                                        @endif
-                                    </td>
                                     <td>{{ $row->name }}</td>
-                                    <td>{{ $row->username }}</td>
-                                    <td>{{ $row->address }}</td>
-                                    <td><a href="mailto:{{ $row->email }}" target="_blank">{{ $row->email }}</a></td>
-                                    <td>{{ $row->phone }}</td>
                                     <td>
-                                        @if($row->privilege)
-                                            <span class="{{ getBadge($row->privileges->color) }}">{{ $row->privileges->name }}</span>
+                                        @if($row->user)
+                                            <span>{{ $row->users->name }}</span>
+                                            <br><span class="{{ getBadge($row->users->privileges->color) }}">{{ $row->users->privileges->name }}</span>
+                                        @else
+                                            <span>-</span>
                                         @endif
                                     </td>
                                     {{--                                    Jika can CRUD maka munculkan tombol--}}
-                                    @if(getUserLevel('users') >= CAN_CRUD)
+                                    @if(getUserLevel('branches') >= CAN_CRUD)
                                         <td class="pe-3 w-2-slot">
-                                            <div class="d-flex">
-                                                @include('form.button.crud', ['url' => 'user/', 'type' => $type, 'id' => $row->id])
+                                            <div class="d-flex justify-content-center">
+                                                @include('form.button.crud', ['url' => 'branch/coordinator/', 'type' => $type, 'id' => $row->id])
                                             </div>
                                         </td>
                                     @endif
@@ -97,9 +84,9 @@
                         @else
                             <tr>
                                 @php
-                                    $colspan = 8;
-                                    if (getUserLevel('users') >= CAN_CRUD) {
-                                        $colspan += 1; // ada bagian untuk action button
+                                    $colspan = 3;
+                                    if (getUserLevel('branches') >= CAN_CRUD) {
+                                        $colspan += 1; // ada baguian untuk action button
                                     }
                                 @endphp
                                 <td class="text-muted text-center" colspan="{{ $colspan }}">Tidak Ada Data</td>